+++ .gitignore
@@ -10,3 +10,4 @@ node_modules/
 npm-debug.log*
 Thumbs.db
 .vscode/
+run.js
